Get the whole thing building and testing in separate modules
