+++ parent-pom.xml
@@ -94,6 +94,12 @@
             <version>2.3.1</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.hamcrest</groupId>
+            <artifactId>hamcrest-core</artifactId>
+            <scope>test</scope>
+            <version>1.1</version>
+        </dependency>
     </dependencies>
 
     <licenses>
