+++ rollbar-payload/src/main/java/com/rollbar/payload/PayloadSerializer.java
@@ -1,4 +1,4 @@
-package com.rollbar.payload.utilities;
+package com.rollbar.payload;
 
 import com.rollbar.payload.Payload;
