+++ rollbar-payload/src/main/java/com/rollbar/payload/data/Level.java
@@ -1,6 +1,6 @@
 package com.rollbar.payload.data;
 
-import com.rollbar.payload.utilities.JsonSerializable;
+import com.rollbar.utilities.JsonSerializable;
 
 /**
  * The Level of a Rollbar Report.
