+++ rollbar-payload/src/main/java/com/rollbar/payload/data/body/Message.java
@@ -1,8 +1,8 @@
 package com.rollbar.payload.data.body;
 
-import com.rollbar.payload.utilities.ArgumentNullException;
-import com.rollbar.payload.utilities.Extensible;
-import com.rollbar.payload.utilities.Validate;
+import com.rollbar.utilities.ArgumentNullException;
+import com.rollbar.utilities.Extensible;
+import com.rollbar.utilities.Validate;
 
 import java.util.HashSet;
 import java.util.Map;
