+++ rollbar-payload/src/test/java/com/rollbar/payload/data/PersonTest.java
@@ -1,10 +1,9 @@
 package com.rollbar.payload.data;
 
-import com.rollbar.GetAndSet;
-import com.rollbar.TestThat;
-import com.rollbar.payload.utilities.ArgumentNullException;
-import com.rollbar.payload.utilities.InvalidLengthException;
-import com.sun.org.apache.xpath.internal.Arg;
+import com.rollbar.testing.GetAndSet;
+import com.rollbar.testing.TestThat;
+import com.rollbar.utilities.ArgumentNullException;
+import com.rollbar.utilities.InvalidLengthException;
 import org.junit.Before;
 import org.junit.Test;
