+++ rollbar-payload/src/test/java/com/rollbar/payload/data/body/CodeContextTest.java
@@ -1,10 +1,8 @@
 package com.rollbar.payload.data.body;
 
-import com.rollbar.GetAndSet;
-import com.rollbar.TestThat;
+import com.rollbar.testing.GetAndSet;
+import com.rollbar.testing.TestThat;
 import org.junit.Test;
-
-import static org.junit.Assert.*;
 
 /**
  * Created by chris on 11/25/15.
