+++ rollbar-payload/src/test/java/com/rollbar/payload/data/body/FrameTest.java
@@ -1,11 +1,10 @@
 package com.rollbar.payload.data.body;
 
-import com.rollbar.GetAndSet;
-import com.rollbar.TestThat;
-import com.rollbar.payload.utilities.ArgumentNullException;
+import com.rollbar.testing.GetAndSet;
+import com.rollbar.testing.TestThat;
+import com.rollbar.utilities.ArgumentNullException;
 import org.junit.Before;
 import org.junit.Test;
-import org.omg.CORBA.OBJECT_NOT_EXIST;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
