+++ rollbar-payload/src/test/java/com/rollbar/payload/data/body/TraceChainTest.java
@@ -1,10 +1,11 @@
 package com.rollbar.payload.data.body;
 
-import com.rollbar.GetAndSet;
-import com.rollbar.TestThat;
+import com.rollbar.testing.*;
+
 import com.rollbar.http.ConnectionFailedException;
-import com.rollbar.payload.utilities.ArgumentNullException;
-import com.rollbar.payload.utilities.InvalidLengthException;
+import com.rollbar.utilities.ArgumentNullException;
+import com.rollbar.utilities.InvalidLengthException;
+
 import org.junit.Before;
 import org.junit.Test;
