+++ rollbar-testing/pom.xml
@@ -2,8 +2,6 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
     <parent>
         <groupId>com.rollbar</groupId>
         <artifactId>rollbar-parent</artifactId>
@@ -11,13 +9,10 @@
         <relativePath>../parent-pom.xml</relativePath>
     </parent>
 
-    <name>rollbar-payload</name>
-    <description>A set of classes that allow reporting errors to Rollbar.</description>
-    <url>http://www.rollbar.com</url>
+    <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.rollbar</groupId>
-    <artifactId>rollbar</artifactId>
-    <version>0.4.0</version>
+    <artifactId>rollbar-testing</artifactId>
+    <version>0.0.1-alpha</version>
 
     <dependencies>
         <dependency>
@@ -25,12 +20,22 @@
             <artifactId>rollbar-utilities</artifactId>
             <version>0.0.1-alpha</version>
         </dependency>
-        <dependency>
-            <groupId>com.rollbar</groupId>
-            <artifactId>rollbar-testing</artifactId>
-            <version>0.0.1-alpha</version>
-            <type>test-jar</type>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <artifactId>maven-jar-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <id>test-jar</id>
+                        <phase>test-compile</phase>
+                        <goals>
+                            <goal>test-jar</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
 </project>
