+++ rollbar-utilities/pom.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>com.rollbar</groupId>
+        <artifactId>rollbar-parent</artifactId>
+        <version>0.0.1-alpha</version>
+        <relativePath>../parent-pom.xml</relativePath>
+    </parent>
+
+    <artifactId>rollbar-utilities</artifactId>
+    <version>0.0.1-alpha</version>
+</project>
